Avoid signed overflow in per-picture QP overrides

+++ encoder/ratecontrol.c
@@ -1553,7 +1553,7 @@ void x264_ratecontrol_start( x264_t *h, int i_force_qp, int overhead )
         }
     }
     if( i_force_qp != X264_QP_AUTO )
-        q = i_force_qp - 1;
+        q = (double)i_force_qp - 1;
 
     q = x264_clip3f( q, h->param.rc.i_qp_min, h->param.rc.i_qp_max );
